minor ref test simplifications
