+++ src/webgpu/web-platform/reftests/gpu_ref_test.ts
@@ -1,49 +1,23 @@
-import { DevicePool } from '../../../common/framework/gpu/implementation.js';
-import { assert, unreachable } from '../../../common/framework/util/util.js';
+import { assert } from '../../../common/framework/util/util.js';
 
-const devicePool = new DevicePool();
+declare function takeScreenshotDelayed(ms: number): void;
 
-export class GPURefTest {
-  private objects: { device: GPUDevice; queue: GPUQueue } | undefined = undefined;
-  initialized = false;
+interface GPURefTest {
+  readonly device: GPUDevice;
+  readonly queue: GPUQueue;
+}
 
-  get device(): GPUDevice {
-    assert(this.objects !== undefined);
-    return this.objects.device;
-  }
+export async function runRefTest(fn: (t: GPURefTest) => Promise<void>): Promise<void> {
+  assert(
+    typeof navigator !== 'undefined' && navigator.gpu !== undefined,
+    'No WebGPU implementation found'
+  );
 
-  get queue(): GPUQueue {
-    assert(this.objects !== undefined);
-    return this.objects.queue;
-  }
+  const adapter = await navigator.gpu.requestAdapter();
+  const device = await adapter.requestDevice();
+  const queue = device.defaultQueue;
 
-  async init(): Promise<void> {
+  await fn({ device, queue });
 
-    const device = await devicePool.acquire();
-    const queue = device.defaultQueue;
-    this.objects = { device, queue };
-
-    try {
-      await device.popErrorScope();
-      unreachable('There was an error scope on the stack at the beginning of the test');
-    } catch (ex) { }
-
-    device.pushErrorScope('out-of-memory');
-    device.pushErrorScope('validation');
-
-    this.initialized = true;
-  }
-
-  async finalize(): Promise<void> {
-    if (this.initialized) {
-      try {
-        await this.device.popErrorScope()
-      } catch (ex) { }
-    }
-
-    // Note: finalize is called even if init was unsuccessful.
-    if (this.objects) {
-      devicePool.release(this.objects.device);
-    }
-  }
+  takeScreenshotDelayed(50);
 }
