+++ src/webgpu/web-platform/reftests/present_to_canvas.ts
@@ -1,23 +1,9 @@
-import { GPURefTest } from './gpu_ref_test.js';
+import { runRefTest } from './gpu_ref_test.js';
 
-async function PresentToCanvasTest(): Promise<void> {
-  const t = new GPURefTest();
-  await t.init();
-  if (typeof document === 'undefined') {
-    // Skip if there is no document (Workers, Node)
-    console.log('DOM is not available to create canvas element');
-    return;
-  }
+runRefTest(async t => {
+  const canvas = document.getElementById('gpucanvas') as HTMLCanvasElement;
 
-  const canvas: any = document.getElementById('gpucanvas');
-  if (canvas === null) {
-    console.log('DOM is null');
-    return;
-  }
-
-  // TODO: fix types so these aren't necessary
-  // tslint:disable-next-line: no-any
-  const ctx: any = canvas.getContext('gpupresent');
+  const ctx = (canvas.getContext('gpupresent') as unknown) as GPUCanvasContext;
   const swapChain = ctx.configureSwapChain({
     device: t.device,
     format: 'rgba8unorm',
@@ -38,8 +24,4 @@ async function PresentToCanvasTest(): Promise<void> {
   });
   pass.endPass();
   t.device.defaultQueue.submit([encoder.finish()]);
-  await t.finalize();
-}
-PresentToCanvasTest().then(() => {
-  takeScreenshotDelayed(50);
-})
+});
